# Perf optimizations with function ID resolution