<!-- Please provide all the information below.  -->

Perf optimizations when resolving function IDs.

For the method optimization alone, these are the high level results:

Case	ResultsGraph	Mean	Min	Max	Range	    AllocatedBytes	Operations	Phase
TestNew		14.90 ns	14.58 ns	15.31 ns	5%	1,544	503,316,480	Complete
TestOld		29.29 ns	28.85 ns	30.14 ns	4%	1,652	234,881,024	Complete


### Pull request checklist

* [x] My changes **do not** require documentation changes
    * [ ] Otherwise: Documentation issue linked to PR
* [x] My changes **should not** be added to the release notes for the next release
    * [ ] Otherwise: I've added my notes to `release_notes.md`
* [x] My changes **do not** need to be backported to a previous version
    * [ ] Otherwise: Backport tracked by issue/PR #issue_or_pr
* [x] My changes **do not** require diagnostic events changes
    * Otherwise: I have added/updated all related diagnostic events and their documentation (Documentation issue linked to PR)
* [x] I have added all required tests (Unit tests, E2E tests)